requirements.txt
Django~=2.1.2
Psycopg2
djangorestframework
Pillow